Support starting local processes on Windows
This allows to start and interact with local processes on Windows. The `process` class constructor is very similar to starting processes on linux, but stripped of unsupported parameters.

PythonForWindows' WinProcess class is used to access the PEB of the new process and read useful information during runtime like imports and cwd.

This is based on https://github.com/masthoon/pwintools :heart: